Predict Restaurant Customer Satisfaction Dataset
This dataset provides comprehensive information on customer visits to restaurants, including demographic details, visit-specific metrics, and customer satisfaction ratings. It is designed to facilitate predictive modeling and analytics in the hospitality industry, focusing on factors that drive customer satisfaction. This dataset is mainly for Predict Restaurant Customer Satisfaction Dataset, there are some contains which included in the dataset.